Restructure template tags
