+++ historie/templates/historie/_modalDataIncludes.html
@@ -1,4 +1,4 @@
-{% load to_class_name %}
+{% load t_historie.to_class_name %}
 
 {% if entry|to_class_name == "HistoricalMitglied" %}
 {% include 'historie/modalData/mitgliedModalData.html' %}
