+++ historie/templates/historie/rowContent.html
@@ -1,4 +1,4 @@
-{% load to_class_name %}
+{% load t_historie.to_class_name %}
 
 <div id="modal{{entry|to_class_name}}{{entry.history_id}}" class="modal">
   <div class="modal-content">
